Declaration: Move `originalType` to be a function
An common issue with the syntax tree is that the coupling between `Declaration`
and `Type` leads to a lot of casting. The prime example is `FunctionDeclaration`:
its `type` member is used all over the place, however it is statically typed as `Type`.
Users holding a `FuncDeclaration` (reasonably) expects the type to be `TypeFunction`,
but any `TypeFunction`-specific usage will require a `cast`.
Fortunately (?), such casts are static, meaning there's no runtime cost involved
(but they are unsafe), however they clutter the source and make type-safe improvements harder.

This is a first step towards fixing that issue: By moving `originalType` to be a function,
we can use covariant overloads in child classes to return an instance of a `TypeFunction`
instead of simply a `Type`. This allows us to remove any cast when the declaration type
is known to be a `FuncDeclaration`. This approach was already used in `syntaxCopy` with
good success, and can soon be extended to the `type` field itself.